Fixed queue editor number displays
Now when a task is removed, it won't count as a task shown.
Now when a task is added back it will count as a task shown.
Now when a task doesn't have a 'task_name' key, it will use the task JSON as a string to compare results from the search bar.