Ability to select stored server routes
Updated addTextInput method in Popup class to accept an array of strings called datalist to create a drop down selection with the options/values in the datalist.
Now when the Add Server popup is generated, if there are locally stored routes they will be options for the user to select in a dropdown.